refactored group list functionality into nested directory
